requirements.txt
pandas
requests
yfinance
numpy
streamlit
ta
plotly
alpha_vantage
python-dotenv

